src/document_completed/requirements.txt
flask
dapr
cloudevents
uvicorn
typing-extensions
azure-storage-blob
pypdf
azure-search-documents==11.4.0b6
azure-ai-formrecognizer==3.3.2
azure-ai-textanalytics
azure-identity
openai[datalib]==0.27.8
tenacity==8.2.2
nanoid